test(deepseek_v4): sequester the tests that need HF as their oracle
The DeepSeek V4 tests use `transformers.models.deepseek_v4` as their correctness oracle, and
that package only exists from transformers 5.15. Wanting it is the only reason this branch
carried a transformers bump, so split each file along the oracle line and invoke the HF half
explicitly instead:

    uv run --with 'transformers==5.15.0' pytest tests/unit/train/models/test_deepseek_v4_hf.py

Tests that compare against HF move to `*_hf.py`, each guarded by `pytest.importorskip` so the
module skips rather than errors under the pin. Tests that only used HF as a source of random
weights stay put and build prime-rl's model directly; `_randomize` now also draws the
`expert_bias` buffer, which prime-rl hangs one level above where HF does and would otherwise
have been left at zero. Shared fixtures move to `deepseek_v4_helpers.py` and
`deepseek_v4_temp_helpers.py`; `test_deepseek_v4_temp.py`'s docstring now describes the file
it has become rather than the scratch file it started as.

Three tests are new, covering what the HF oracle used to and no longer can by default:
`layer_types` is rejected by prime-rl's own validator (which until now never ran, since the
base one raised first), the legacy `compress_ratios` translation against written-out
schedules, and prime-rl's on-disk key set against the names vLLM's weights mapper produces.
That last one replaces an HF-derived key-set check, and pins the mapped names rather than
asking the mapper to reject anything: it returns unrecognized keys unchanged, never `None`,
so a "nothing was dropped" assertion would hold for any input at all.
